Change default for scroll to false
A large number of clients have requested the Add to Cart functionality does not scroll to top. The behavior
of scrolling should be an option, but not the default one since it goes against the purpose of this extension,
which is to let the user add to their cart without disrupting the browsing experience. When it scroll to top,
it's very much like refreshing the page.